fix: generate types only for FE
