+++ fe/lib/types/seek.graphql.ts
