+++ fe/lib/types/seekFragmentTypes.ts
@@ -1,1 +1,53 @@
-export * from '../../../types/seekFragmentTypes';
+
+      export interface IntrospectionResultData {
+        __schema: {
+          types: {
+            kind: string;
+            name: string;
+            possibleTypes: {
+              name: string;
+            }[];
+          }[];
+        };
+      }
+      const result: IntrospectionResultData = {
+  "__schema": {
+    "types": [
+      {
+        "kind": "INTERFACE",
+        "name": "ApplicationQuestionnaireComponent",
+        "possibleTypes": [
+          {
+            "name": "ApplicationQuestion"
+          },
+          {
+            "name": "ApplicationPrivacyConsent"
+          }
+        ]
+      },
+      {
+        "kind": "INTERFACE",
+        "name": "ApplicationQuestionnaireComponentResponse",
+        "possibleTypes": [
+          {
+            "name": "ApplicationQuestionResponse"
+          },
+          {
+            "name": "ApplicationPrivacyConsentResponse"
+          }
+        ]
+      },
+      {
+        "kind": "INTERFACE",
+        "name": "Event",
+        "possibleTypes": [
+          {
+            "name": "CandidateApplicationCreatedEvent"
+          }
+        ]
+      }
+    ]
+  }
+};
+      export default result;
+
