--- fe/lib/types/seekTypes.ts
@@ -1,1 +1,0 @@
-export * from '../../../types/seek.graphql';
